Added phrase listener and command
+ fixes

+++ app/src/listeners/SabatuListener.ts
@@ -1,13 +1,15 @@
 import { Client } from "discord.js";
-import { console } from '../util/log';
+import { console } from "../util/log";
 
 export default async function (client: Client) {
-	client.on('message', async (msg) => {
-		if (msg.content.toLowerCase().includes('sabatu')) {
-			const msg2 = await msg.channel.send('🙂');
-			msg2.delete();
-		}
-	});
+  client.on("message", async (msg) => {
+    if (msg.author.bot) return;
 
-	console.info('Initialized Sabatu listener');
+    if (msg.content.toLowerCase().includes("sabatu")) {
+      const msg2 = await msg.channel.send("🙂");
+      msg2.delete();
+    }
+  });
+
+  console.info("Initialized Sabatu listener");
 }
