+++ app/src/listeners/SpelosListener.ts
@@ -1,18 +1,20 @@
 import { Client } from "discord.js";
-import { console } from '../util/log';
+import { console } from "../util/log";
 
 export default async function (client: Client) {
-	client.on('message', async (msg) => {
-		if (msg.content.toLowerCase().includes('spelos')) {
-			if (Math.random() < 0.5) {
-				const msg2 = await msg.channel.send('beg for it :)');
-				msg2.delete();
-			} else {
-				const msg2 = await msg.channel.send('🤗');
-				msg2.delete();
-			}
-		}
-	});
+  client.on("message", async (msg) => {
+    if (msg.author.bot) return;
 
-	console.info('Initialized Spelos listener');
+    if (msg.content.toLowerCase().includes("spelos")) {
+      if (Math.random() < 0.5) {
+        const msg2 = await msg.channel.send("beg for it :)");
+        msg2.delete();
+      } else {
+        const msg2 = await msg.channel.send("🤗");
+        msg2.delete();
+      }
+    }
+  });
+
+  console.info("Initialized Spelos listener");
 }
